# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Реактив CRM",
    "description": "CRM продаж и закупок химической продукции. Настраиваемые финансовые профили, права и шаблоны. Количества и деньги — десятичные строки.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analytics/dashboard": {
      "get": {
        "tags": [
          "Аналитика"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_v1_analytics_dashboard_get",
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "owner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "name": "include_test",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Test"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Dashboard Api V1 Analytics Dashboard Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/analytics/requests": {
      "get": {
        "tags": [
          "Аналитика"
        ],
        "summary": "Drilldown",
        "operationId": "drilldown_api_v1_analytics_requests_get",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "new",
              "title": "Metric"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "owner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "name": "include_test",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Test"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Drilldown Api V1 Analytics Requests Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/analytics/invoices": {
      "get": {
        "tags": [
          "Аналитика"
        ],
        "summary": "Invoices",
        "operationId": "invoices_api_v1_analytics_invoices_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Invoices Api V1 Analytics Invoices Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/analytics/export.xlsx": {
      "get": {
        "tags": [
          "Аналитика"
        ],
        "summary": "Export",
        "operationId": "export_api_v1_analytics_export_xlsx_get",
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "owner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "name": "include_test",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Test"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
...
  },
...
}